fix(state): extract state-key enums so vscode-free zones can use them
Codex P1 (texFormatter.ts:3) on #3185: importing WorkspaceStateKey
from @common/state in src/latex/ violates the vscode-free-zone rule
in CLAUDE.md/AGENTS.md, because @common/state resolves to
stateManager.ts which imports vscode. The enum is pure data; only the
runtime managers (workspaceSM, globalSM, initializeStateManagers)
need vscode.

Move WorkspaceStateKey, GlobalStateKey, and INSTRUCTION_PREFIX to a
new src/common/state/stateKeys.ts that has zero vscode coupling.
stateManager.ts re-exports them so existing direct imports keep
working unchanged. The barrel (@common/state) still works for vscode-
allowed zones (commands/, frontend/, etc.) — they pull in the runtime
managers anyway.

Switch the five vscode-free-zone consumers added in chunk 2b to import
from the new path:
- src/agent/output/compileCheck.ts
- src/agent/output/LatexDiffManager.ts
- src/latex/latexdiff.ts
- src/latex/latexdiff/diffCommandExecutor.ts
- src/latex/texFormatter.ts

(Pre-existing @common/state imports in src/agent/runtime/* etc. share
the same architectural concern but predate this PR; cleaning them up
is a separate effort outside this feature.)

Typecheck and fast build are clean.
